+++ lib/oli/authoring/project_export_worker.ex
@@ -9,6 +9,7 @@ defmodule Oli.Authoring.ProjectExportWorker do
   alias Oli.Utils
   alias Oli.Authoring.Broadcaster
   alias Oli.Authoring.Course
+  alias Oli.Authoring.Course.Project
 
   @impl Oban.Worker
   def perform(%Oban.Job{args: %{"project_slug" => project_slug} = _args}) do
@@ -56,5 +57,41 @@ defmodule Oli.Authoring.ProjectExportWorker do
     Course.update_project_latest_export_url(project_slug, full_upload_url, timestamp)
 
     {full_upload_url, timestamp}
+  end
+
+  @doc """
+  Generates a project export for the given project if one is not already in progress
+  """
+  def generate_project_export(project) do
+    case project_export_status(project) do
+      {:in_progress} ->
+        {:error, "Project export is already in progress"}
+
+      _ ->
+        %{project_slug: project.slug}
+        |> Oli.Authoring.ProjectExportWorker.new()
+        |> Oban.insert()
+    end
+  end
+
+  def project_export_status(project) do
+    if Course.export_in_progress?(project.slug, "project_export") do
+      # export is in progress
+      {:in_progress}
+    else
+      case project do
+        # export is created and completed
+        %Project{
+          latest_export_url: export_url,
+          latest_export_timestamp: export_timestamp
+        }
+        when not is_nil(export_url) and not is_nil(export_timestamp) ->
+          {:available, export_url, export_timestamp}
+
+        # export has not been created yet
+        _ ->
+          {:not_available}
+      end
+    end
   end
 end
