+++ lib/oli_web/components/project/async_exporter.ex
@@ -184,12 +184,12 @@ defmodule OliWeb.Components.Project.AsyncExporter do
     ~H"""
     <%= case @project_export_status do %>
       <% status when status in [:not_available, :expired] -> %>
-        <.button variant={:link} phx-click={@on_generate_project_export}>
+        <.button variant={:link} class="!px-3" phx-click={@on_generate_project_export}>
           Export
         </.button>
         <div>Download this project and its contents</div>
       <% :in_progress -> %>
-        <.button variant={:link} disabled>Export</.button>
+        <.button variant={:link} class="!px-3" disabled>Export in Progress</.button>
         <div class="flex flex-col">
           <div>Download this project and its contents</div>
           <div class="text-sm text-gray-500">
@@ -198,7 +198,7 @@ defmodule OliWeb.Components.Project.AsyncExporter do
           </div>
         </div>
       <% :available -> %>
-        <.button variant={:link} href={@project_export_url} download>
+        <.button variant={:link} class="!px-3" href={@project_export_url} download>
           <i class="fa-solid fa-download mr-1"></i> Download Latest Export
         </.button>
         <div class="flex flex-col">
@@ -211,7 +211,7 @@ defmodule OliWeb.Components.Project.AsyncExporter do
           </div>
         </div>
       <% :error -> %>
-        <.button variant={:link} phx-click={@on_generate_project_export}>
+        <.button variant={:link} class="!px-3" phx-click={@on_generate_project_export}>
           Export
         </.button>
         <div class="flex flex-col">
